Refactoring by removing unused code.
